+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Option to remove R seed from distance computation
 - `most_variable_features` method
 - `characteristic_features` method
+- `most_correlated` method
 
 ## v1.0.x - unreleased
 ### Changed